requirements.txt
datasets
pandas
google-genai
tqdm
python-dotenv
